+++ addons/primitives/high_primitives.c
@@ -133,7 +133,7 @@ void al_draw_triangle(float x1, float y1, float x2, float y2,
       /*
        * If the triangle is very flat, draw it as a line
        */
-      if(fabs(cross) < 0.0001f) {
+      if(fabsf(cross) < 0.0001f) {
          float tx, ty, lx, ly;
          float len;
          /*
@@ -504,7 +504,7 @@ void al_draw_pieslice(float cx, float cy, float r, float start_theta,
    }
    
    if (thickness <= 0) {
-      num_segments = fabs(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf(r));
+      num_segments = fabsf(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf(r));
 
       if (num_segments < 2)
          num_segments = 2;
@@ -533,7 +533,7 @@ void al_draw_pieslice(float cx, float cy, float r, float start_theta,
       float midpoint_dir_y = sinf(midangle);
       float side_dir_x = cosf(start_theta);
       float side_dir_y = sinf(start_theta);
-      float sine_half_delta = fabs(side_dir_x * midpoint_dir_y - side_dir_y * midpoint_dir_x); /* Cross product */
+      float sine_half_delta = fabsf(side_dir_x * midpoint_dir_y - side_dir_y * midpoint_dir_x); /* Cross product */
       float connect_len = ht / sine_half_delta;
       bool blunt_tip = connect_len > 2 * thickness;
       
@@ -653,7 +653,7 @@ void al_draw_filled_pieslice(float cx, float cy, float r, float start_theta,
    
    ASSERT(r >= 0);
    
-   num_segments = fabs(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf(r));
+   num_segments = fabsf(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf(r));
 
    if (num_segments < 2)
       num_segments = 2;
@@ -785,7 +785,7 @@ void al_draw_elliptical_arc(float cx, float cy, float rx, float ry, float start_
 
    ASSERT(rx >= 0 && ry >= 0);
    if (thickness > 0) {
-      int num_segments = fabs(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf((rx + ry) / 2.0f));
+      int num_segments = fabsf(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf((rx + ry) / 2.0f));
       int ii;
 
       if (num_segments < 2)
@@ -804,7 +804,7 @@ void al_draw_elliptical_arc(float cx, float cy, float rx, float ry, float start_
       
       al_draw_prim(vertex_cache, 0, 0, 0, 2 * num_segments, ALLEGRO_PRIM_TRIANGLE_STRIP);
    } else {
-      int num_segments = fabs(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf((rx + ry) / 2.0f));
+      int num_segments = fabsf(delta_theta / (2 * ALLEGRO_PI) * ALLEGRO_PRIM_QUALITY * scale * sqrtf((rx + ry) / 2.0f));
       int ii;
 
       if (num_segments < 2)
